+++ src/rook/io/__init__.py
@@ -1,0 +1,5 @@
+"""Dataset input and output utilities."""
+
+from .datasets import open_dataset
+
+__all__ = ["open_dataset"]
